+++ analytics/upress_reporting/models/reports/views.py
@@ -1,6 +1,5 @@
 import csv
 
-from logger import createLog
 from models.aggregators.view_data_aggregator import ViewDataAggregator
 from models.reports.counter_5_report import Counter5Report
 
@@ -8,13 +7,11 @@ from models.reports.counter_5_report import Counter5Report
 class ViewsReport(Counter5Report):
     def __init__(self, *args):
         super().__init__(*args)
-        self.view_request_parser = ViewDataAggregator(
-            self.publisher, self.pandas_date_range)
-        self.logger = createLog("views_report")
+        self.view_request_parser = ViewDataAggregator(self.publisher, self.pandas_date_range)
 
     def build_report(self):
         # TODO: move to general counter 5 class?
-        self.logger.info("Building views report...")
+        print("Building views report...")
 
         header = self.build_header()
         view_events = self.view_request_parser.events
@@ -33,9 +30,9 @@ class ViewsReport(Counter5Report):
                 for title in final_data:
                     writer.writerow(title.values())
 
-            self.logger.info("Views report generation complete!")
+            print("Views report generation complete!")
         else:
-            self.logger.info("No view events found in reporting period!")
+            print("No view events found in reporting period!")
 
     def build_header(self):
         return {